tweetdi extractor functions completed
